requirements.txt
Django~=5.1.3
drf-spectacular~=0.28.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
gunicorn~=23.0.0
Pillow~=11.0.0
psycopg2-binary~=2.9.10
social-auth-app-django~=5.4.2
social-auth-core~=4.5.4
